Move filter query construction to api layer
