+++ src/api/instances/query/get.ts
@@ -6,6 +6,43 @@ import { instanceKeys } from "..";
 import { useApiKey } from "~/util/store/apiKey";
 import { useNamespace } from "~/util/store/namespace";
 
+export type FiltersObj = {
+  AS?: { type: "CONTAINS"; value: string };
+  STATUS?: {
+    type: "MATCH";
+    value: "pending" | "complete" | "cancelled" | "failed";
+  };
+  TRIGGER?: {
+    type: "MATCH";
+    value: "api" | "cloudevent" | "instance" | "cron";
+  };
+  // TODO: use Date type (but display components need a string value)
+  AFTER?: { type: "AFTER"; value: string };
+  BEFORE?: { type: "BEFORE"; value: string };
+};
+
+const getFilterQuery = (filters: FiltersObj) => {
+  let query = "";
+  const filterFields = Object.keys(filters) as Array<keyof FiltersObj>;
+
+  filterFields.forEach((field) => {
+    const filterItem = filters[field];
+    // guard needed because TS thinks filterItem may be undefined
+    if (!filterItem) {
+      return;
+    }
+
+    // FilterFields uses BEFORE and AFTER as distinct types. In the query,
+    // the format is filter.field=CREATED&filter.type=BEFORE|AFTER
+    const queryField = ["AFTER", "BEFORE"].includes(field) ? "CREATED" : field;
+
+    query = query.concat(
+      `&filter.field=${queryField}&filter.type=${filterItem.type}&filter.val=${filterItem.value}`
+    );
+  });
+  return query;
+};
+
 export const getInstances = apiFactory({
   url: ({
     namespace,
@@ -40,11 +77,11 @@ const fetchInstances = async ({
 export const useInstances = ({
   limit,
   offset,
-  filter,
+  filters,
 }: {
   limit: number;
   offset: number;
-  filter: string;
+  filters: FiltersObj;
 }) => {
   const apiKey = useApiKey();
   const namespace = useNamespace();
@@ -58,7 +95,7 @@ export const useInstances = ({
       apiKey: apiKey ?? undefined,
       limit,
       offset,
-      filter,
+      filter: getFilterQuery(filters),
     }),
     queryFn: fetchInstances,
     enabled: !!namespace,
